add auth

+++ app/services/properties_service.py
@@ -1,11 +1,11 @@
-from app.models.property import Property
+from app.database.models.property import Property
 from app.repositories.properties_repository import PropertiesRepository
 from app.repositories.users_repository import UsersRepository
 from app.common.exceptions import NotFoundError,BadRequestError,ConflictError
 from app.repositories.properties_repository import PropertiesRepository
 from app.validation.property_validation import PropertyValidation
-from app.models.users import Role
-from app.db.session import get_session
+from app.database.models.users import Role
+from app.database.db.session import get_session
 
 class PropertiesService:
     def __init__(self):
@@ -16,8 +16,18 @@ class PropertiesService:
     def create(self, **data) -> Property:
         print("\033[91mCreating property with data:\033[0m", self.users.get(data["owner_id"]))
         
-        PropertyValidation._check_owner(self,data)
-        PropertyValidation.check_property_already_exists(self,data)
+        owner=self.users.get(data["owner_id"])
+     
+        if not owner:
+         raise NotFoundError(f"Owner not found")
+        
+        existing = self.repo.find_by_address_and_unit(data["address"], data["unit_number"])
+        
+        if existing:
+            if existing.owner_id != owner.id:
+                raise ConflictError("This property already exists and is owned by another user.")
+            else:
+                raise ConflictError("You already own this property.")
             
         prop = Property(**data)
         self.repo.create(prop)
@@ -25,7 +35,10 @@ class PropertiesService:
     
 
     def get(self, prop_id: int):
-        PropertyValidation.check_property_exists(self,prop_id)
+        prop = self.repo.get(prop_id)
+        if not prop:
+            raise NotFoundError(f"Property {prop_id} not found in validation")    
+        
         prop = self.repo.get(prop_id)
         return prop
 
@@ -54,11 +67,17 @@ class PropertiesService:
         return updated_prop
 
 
-    def delete(self, prop_id: int) -> None:
-        PropertyValidation.check_property_exists(self,prop_id)
+    def delete(self, prop_id: int, user) -> None:
+        # PropertyValidation.check_property_exists(self,prop_id)
         
         prop = self.repo.get(prop_id)
+        if not prop:
+         raise NotFoundError(f"Property {prop_id} not found")
+    
+        if prop.owner_id != user.id:
+         raise BadRequestError("You do not have permission to delete this property")
+
         self.repo.delete(prop)
-        return ({"message": "Property deleted"})
+        return {"message": "Property deleted"}
